chapter06/src/requirements.txt
boto3==1.35.71
pillow==11.0.0
